Move tests for the StudySummary class to test_study.py and fix all structs in redis.py.
